mostly avoid stupid globs expansions

+++ scripts/build.sh
@@ -4,15 +4,15 @@ set -e
 
 # Find the directory we exist within
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
-cd ${DIR}/..
+cd "${DIR}/.."
 
 BUILDDIR=$(pwd)/build
 
 # Make dir
-mkdir -p $BUILDDIR
+mkdir -p "$BUILDDIR"
 
 # Clean build bin dir
-rm -rf $BUILDDIR/*
+rm -rf "${BUILDDIR:?}/*"
 
 function fail () {
 	echo "Aborting due to failure." >&2
@@ -22,9 +22,9 @@ function fail () {
 # Build binary
 cd src/cmd
 for bin in *; do
-  cd $bin
+  cd "$bin"
   set -x
-  go build -o $BUILDDIR/$bin || fail
+  go build -o "$BUILDDIR/$bin" || fail
   set +x
   cd -
 done
